Reject the tile dimensions in the function every binding calls
The tile dimensions are rejected by the PostgreSQL wrapper, while
raster_tile_value_quadbin takes them as uint16_t, so the guard covers
the SQL surface and no other caller. A binding passing 65538 still
samples a two pixel wide tile, and one passing -1 a 65535 pixel wide
one.

Take the dimensions in the type the SQL surface uses, int32, and test
the range in the MEOS function before the narrowing to the tile fields,
as the sibling raquet_make does. The wrapper passes them through, the
error text is the one it raised, and every binding inherits the
rejection.

Cover it from a MEOS caller, which is where the guard was absent.

+++ meos/include/meos_raster.h
@@ -173,7 +173,7 @@ extern int araster_value_gdal(const Temporal *traj, const char *path, int band,
   const Span *vspan);
 
 extern Temporal *raster_tile_value_quadbin(const Temporal *traj,
-  const uint8_t *pixels, size_t pixels_size, uint16_t width, uint16_t height,
+  const uint8_t *pixels, size_t pixels_size, int32 width, int32 height,
   uint64 quadbin, MeosPixType pixtype, double nodata, bool has_nodata);
 
 extern Temporal *raster_tile_value(const Temporal *traj, const Raquet *rq);
